refactor(ui): split the internal transfer screen into Components
`InternalTransferScreen.swift` was 592 lines, and only about half of them were
about the screen. Two views in it were already shared — `TransferSourceRow` is
used by the Send screen and the identities list, `TransferAmountValidationNote`
by the Send screen — so two other screens were reaching into this one's file
for them.

Four pieces move to `InternalTransfer/Components/`:

- `TransferSourceRow` and `TransferAmountValidationNote`, unchanged, now where
  their other callers can find them.
- `TransferPreview`, which only ever needed the formatted amount.
- `TransferEndpointCards`, the From/To cluster: three layouts, the cards they
  are built from, the picker sheet and the endpoint state. Which layout applies
  still follows from `sendFrom` / `receiveInto` alone, so the decision travels
  with the drawing instead of sitting in the screen.

The screen keeps what is its own — the header, amount row, keypad, confirmation
and the amount/unit glue to the view model — and is 283 lines.
