+++ DashWallet/Sources/UI/Payments/InternalTransfer/InternalTransferScreen.swift
@@ -48,19 +48,6 @@ struct InternalTransferScreen: View {
 
     @State private var confirmation: InternalTransferConfirmation?
 
-    /// Standalone screen: which endpoint's balance picker is presented as
-    /// the bottom sheet. `nil` = none — the form shows just the two
-    /// selected-endpoint cards, never all six rows at once. Presentation
-    /// state only — the selection itself lives in the view model.
-    @State private var endpointPicker: EndpointGroup?
-
-    private enum EndpointGroup: String, Identifiable {
-        case from
-        case to
-
-        var id: String { rawValue }
-    }
-
     var body: some View {
         VStack(spacing: 0) {
             if showsHeader {
@@ -81,7 +68,10 @@ struct InternalTransferScreen: View {
                         .padding(.horizontal, 20)
                         .padding(.top, showsHeader ? 12 : 0)
 
-                    directionCards
+                    TransferEndpointCards(
+                        viewModel: viewModel,
+                        sendFrom: sendFrom,
+                        receiveInto: receiveInto)
                         .padding(.horizontal, 20)
 
                     if viewModel.isBlockedBySync {
@@ -95,7 +85,7 @@ struct InternalTransferScreen: View {
                     }
 
                     if viewModel.canContinue {
-                        transferPreview
+                        TransferPreview(amountFormatted: viewModel.dashAmountFormatted)
                             .padding(.horizontal, 20)
                     }
                 }
@@ -195,223 +185,35 @@ struct InternalTransferScreen: View {
 
     // MARK: - From / To cards
 
-    @ViewBuilder
-    private var directionCards: some View {
-        if let source = sendFrom {
-            sendCards(source: source)
-        } else if let target = receiveInto {
-            receiveCards(target: target)
-        } else {
-            swappableCards
-        }
-    }
 
-    /// Send-sheet layout: the source stays pinned as the top card; the rows
-    /// below pick the destination among the other two balances. No swap
-    /// badge — the source is fixed by the tapped balance row.
-    @ViewBuilder
-    private func sendCards(source: ChainNetwork) -> some View {
-        VStack(spacing: 12) {
-            pinnedCard(source, caption: NSLocalizedString("From", comment: ""))
-            selectionGroup(
-                caption: NSLocalizedString("To", comment: ""),
-                networks: availableTargets(for: source),
-                selected: viewModel.resolvedSendTarget,
-                onSelect: viewModel.selectSendTarget)
-        }
-    }
 
-    /// Non-tappable pinned endpoint card (the fixed From of the send sheet).
-    private func pinnedCard(_ network: ChainNetwork, caption: String) -> some View {
-        let display = networkDisplay(network)
-        return sourceRow(
-            iconSystemName: display.icon,
-            caption: caption,
-            title: display.title,
-            balanceTrailing: dashBalanceTrailing(display.balance),
-            selected: false,
-            showsRadio: false,
-            action: {})
-    }
 
-    /// Icon / title / formatted balance for a balance row, one source of
-    /// truth for the pinned cards and picker rows.
-    private func networkDisplay(_ network: ChainNetwork) -> (icon: String, title: String, balance: String) {
-        switch network {
-        case .core:
-            return ("d.circle.fill", network.balanceName, viewModel.coreBalanceFormatted)
-        case .platform:
-            return ("creditcard.fill", network.balanceName, viewModel.platformCreditsFormatted)
-        case .shielded:
-            return ("shield.fill", network.balanceName, viewModel.shieldedBalanceFormatted)
-        }
-    }
 
-    private var swappableCards: some View {
-        VStack(spacing: 12) {
-            collapsedEndpointCard(viewModel.source, caption: NSLocalizedString("From", comment: "")) {
-                presentEndpointPicker(.from)
-            }
-            collapsedEndpointCard(viewModel.resolvedSendTarget, caption: NSLocalizedString("To", comment: "")) {
-                presentEndpointPicker(.to)
-            }
-        }
-        .sheet(item: $endpointPicker) { group in
-            endpointPickerSheet(for: group)
-                .presentationDetents([.height(400)])
-                .presentationDragIndicator(.visible)
-        }
-    }
 
-    private func presentEndpointPicker(_ group: EndpointGroup) {
-        endpointPicker = group
-    }
 
-    /// Bottom-sheet balance picker for one endpoint, sized to slide over
-    /// the keypad area. Every pick applies and dismisses — picking the
-    /// balance already on the opposite side moves that side to its default
-    /// (the view model keeps the endpoints distinct).
-    private func endpointPickerSheet(for group: EndpointGroup) -> some View {
-        let isFrom = group == .from
-        return VStack(alignment: .leading, spacing: 16) {
-            Text(isFrom
-                ? NSLocalizedString("Transfer from", comment: "Internal transfer source picker title")
-                : NSLocalizedString("Transfer to", comment: "Internal transfer destination picker title"))
-                .font(.system(size: 18, weight: .semibold))
-                .foregroundColor(.dash.primaryText)
 
-            selectionGroup(
-                caption: isFrom
-                    ? NSLocalizedString("From", comment: "")
-                    : NSLocalizedString("To", comment: ""),
-                networks: ChainNetwork.allCases,
-                selected: isFrom ? viewModel.source : viewModel.resolvedSendTarget,
-                onSelect: { network in
-                    if isFrom {
-                        viewModel.selectStandaloneSource(network)
-                    } else {
-                        viewModel.selectStandaloneTarget(network)
-                    }
-                    endpointPicker = nil
-                })
 
-            Spacer(minLength: 0)
-        }
-        .padding(.horizontal, 20)
-        .padding(.top, 24)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(Color.dash.primaryBackground)
-    }
 
-    private func collapsedEndpointCard(
-        _ network: ChainNetwork,
-        caption: String,
-        action: @escaping () -> Void
-    ) -> some View {
-        let display = networkDisplay(network)
-        return TransferSourceRow(
-            iconSystemName: display.icon,
-            caption: caption,
-            title: display.title,
-            balanceTrailing: dashBalanceTrailing(display.balance),
-            selected: false,
-            showsRadio: false,
-            showsChevron: true,
-            action: action)
-    }
 
-    /// Receive-sheet layout: the destination stays pinned as the bottom
-    /// card; the rows above pick the source among the other two balances.
-    /// No swap badge — the destination is fixed by the tapped balance row.
-    @ViewBuilder
-    private func receiveCards(target: ChainNetwork) -> some View {
-        VStack(spacing: 12) {
-            selectionGroup(
-                caption: NSLocalizedString("From", comment: ""),
-                networks: availableSources(for: target),
-                selected: viewModel.resolvedReceiveSource,
-                onSelect: viewModel.selectReceiveSource)
 
-            pinnedCard(target, caption: NSLocalizedString("To", comment: ""))
-        }
-    }
 
-    /// Trailing balance amount + Dash currency glyph, shared by every card.
-    private func dashBalanceTrailing(_ formatted: String) -> AnyView {
-        TransferSourceRow.dashBalanceTrailing(formatted)
-    }
 
-    private func selectionGroup(
-        caption: String,
-        networks: [ChainNetwork],
-        selected: ChainNetwork,
-        onSelect: @escaping (ChainNetwork) -> Void
-    ) -> some View {
-        VStack(spacing: 8) {
-            ForEach(networks, id: \.self) { network in
-                let display = networkDisplay(network)
-                sourceRow(
-                    iconSystemName: display.icon,
-                    caption: caption,
-                    title: display.title,
-                    balanceTrailing: dashBalanceTrailing(display.balance),
-                    selected: selected == network,
-                    action: { onSelect(network) })
-            }
-        }
-    }
 
-    /// Tappable source row with a trailing radio indicator — rendered by the
-    /// shared `TransferSourceRow` (also used by the Send screen's From picker).
-    private func sourceRow(
-        iconSystemName: String,
-        caption: String,
-        title: String,
-        balanceTrailing: AnyView,
-        selected: Bool,
-        showsRadio: Bool = true,
-        action: @escaping () -> Void
-    ) -> some View {
-        TransferSourceRow(
-            iconSystemName: iconSystemName,
-            caption: caption,
-            title: title,
-            balanceTrailing: balanceTrailing,
-            selected: selected,
-            showsRadio: showsRadio,
-            action: action)
-    }
 
-    /// Pinned-sheet picker lists (the balance-row arrow sheets): the fixed
-    /// endpoint's balance is left out entirely — only the standalone screen
-    /// shows all three with same-balance taps rejected.
-    private func availableTargets(for source: ChainNetwork) -> [ChainNetwork] {
-        ChainNetwork.allCases.filter { $0 != source }
-    }
 
-    private func availableSources(for target: ChainNetwork) -> [ChainNetwork] {
-        ChainNetwork.allCases.filter { $0 != target }
-    }
 
-    // MARK: - Transfer preview
 
-    private var transferPreview: some View {
-        VStack(spacing: 2) {
-            Text(NSLocalizedString("You will transfer", comment: ""))
-                .font(.system(size: 12))
-                .foregroundColor(Color.dash.secondaryText)
-            HStack(spacing: 4) {
-                Text("~ \(viewModel.dashAmountFormatted)")
-                    .font(.system(size: 13, weight: .semibold))
-                    .foregroundColor(.dash.primaryText)
-                Image("icon_dash_currency")
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(width: 13, height: 13)
-            }
-        }
-        .frame(maxWidth: .infinity)
-    }
+
+
+
+
+
+
+
+
+
+
+
 
     // MARK: - Helpers
 
@@ -476,117 +278,6 @@ struct InternalTransferScreen: View {
     }
 }
 
-// MARK: - TransferAmountValidationNote
 
-/// Inline explanation for a route-specific amount that cannot be submitted.
-/// Keeping this next to the amount/source controls prevents a protective SDK
-/// build-time refusal from becoming the user's first feedback.
-struct TransferAmountValidationNote: View {
-    let message: String
 
-    var body: some View {
-        HStack(alignment: .firstTextBaseline, spacing: 8) {
-            Image(systemName: "exclamationmark.triangle.fill")
-                .font(.system(size: 13))
-                .foregroundColor(.orange)
-            Text(message)
-                .font(.system(size: 13))
-                .foregroundColor(.dash.secondaryText)
-                .fixedSize(horizontal: false, vertical: true)
-            Spacer(minLength: 0)
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
 
-// MARK: - TransferSourceRow
-
-/// Tappable balance row with icon, caption ("From"/"To"), title, trailing
-/// balance, and a radio indicator — shared by the internal transfer screen's
-/// source picker and the Send screen's From picker.
-struct TransferSourceRow: View {
-    let iconSystemName: String
-    let caption: String
-    let title: String
-    let balanceTrailing: AnyView
-    let selected: Bool
-    /// False renders a fixed (non-picker) endpoint card: no radio circle
-    /// and no selection border.
-    var showsRadio: Bool = true
-    /// With `showsRadio` false: true renders a tappable collapsed selector
-    /// (chevron trailing, expands a picker on tap) instead of a fixed card.
-    var showsChevron: Bool = false
-    var action: () -> Void
-
-    /// Trailing balance amount + Dash currency glyph, the standard trailing
-    /// content for these rows.
-    static func dashBalanceTrailing(_ formatted: String) -> AnyView {
-        AnyView(
-            HStack(spacing: 2) {
-                Text(formatted)
-                    .font(.system(size: 14, weight: .semibold))
-                    .foregroundColor(.dash.primaryText)
-                Image("icon_dash_currency")
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(width: 14, height: 14)
-            })
-    }
-
-    var body: some View {
-        Button(action: action) {
-            HStack(spacing: 14) {
-                Image(systemName: iconSystemName)
-                    .font(.system(size: 18, weight: .semibold))
-                    .foregroundColor(.blue)
-                    .frame(width: 36, height: 36)
-                    .background(Color.dash.blue.opacity(0.08))
-                    .clipShape(Circle())
-
-                VStack(alignment: .leading, spacing: 2) {
-                    Text(caption)
-                        .font(.system(size: 11))
-                        .foregroundColor(Color.dash.secondaryText)
-                    Text(title)
-                        .font(.system(size: 15, weight: .semibold))
-                        .foregroundColor(.dash.primaryText)
-                }
-
-                Spacer()
-
-                balanceTrailing
-
-                if showsRadio {
-                    radioIndicator
-                } else if showsChevron {
-                    Image(systemName: "chevron.down")
-                        .font(.system(size: 13, weight: .semibold))
-                        .foregroundColor(Color.dash.secondaryText)
-                }
-            }
-            .padding(.horizontal, 14)
-            .padding(.vertical, 12)
-            .background(Color.dash.secondaryBackground)
-            .overlay(
-                RoundedRectangle(cornerRadius: 12)
-                    .stroke(showsRadio && selected ? Color.dash.blue : Color.clear,
-                            lineWidth: showsRadio && selected ? 1.5 : 0))
-            .cornerRadius(12)
-        }
-        .buttonStyle(.plain)
-        .disabled(!showsRadio && !showsChevron)
-    }
-
-    private var radioIndicator: some View {
-        ZStack {
-            Circle()
-                .stroke(selected ? Color.dash.blue : Color.dash.gray300.opacity(0.6), lineWidth: 1.5)
-                .frame(width: 18, height: 18)
-            if selected {
-                Circle()
-                    .fill(Color.dash.blue)
-                    .frame(width: 10, height: 10)
-            }
-        }
-    }
-}
